Generate random doubles not floats.

+++ benchmarks/zisa/math/poly2d.cpp
@@ -6,7 +6,7 @@
 static void bm_poly2d_eval(benchmark::State &state) {
   std::random_device rd;
   std::mt19937 e(rd());
-  std::uniform_real_distribution<float> dist(0, 1.0);
+  std::uniform_real_distribution<double> dist(0, 1.0);
 
   auto rand = [&e, &dist]() { return dist(e); };
